requirements.txt
accelerate==0.19.0
clean-fid==0.1.35
einops==0.6.1
numpy==1.24.3
Pillow==9.5.0
PyYAML==6.0
tqdm==4.65.0
wandb==0.15.0
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1+cu117
torchvision==0.14.1+cu117